requirements.txt
Django==2.2.3
django-macaddress==1.5.0
djangorestframework==3.9.3
